[API] Extract shop user logic rules for orders to separate extension
